src/main/java/receipe/receipeproject/converters/NotesCommandToNotes.java
package receipe.receipeproject.converters;

import com.sun.istack.Nullable;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;
import receipe.receipeproject.commands.NotesCommand;
import receipe.receipeproject.domain.Notes;

@Component
public class NotesCommandToNotes implements Converter<NotesCommand, Notes> {
    @Synchronized
    @Nullable
    @Override
    public Notes convert(NotesCommand source) {
        if (source == null) {
            return null;
        }
        final Notes notes = new Notes();
        notes.setId(source.getId());
        notes.setRecipesNotes(source.getRecipeNotes());
        return notes;
    }
}
